add get_train_transforms to shared
