+++ demo/lightning/image-segmentation/dataset.py
@@ -15,12 +15,11 @@
  """
 
 import io
-import random
 import numpy as np
-import scipy.ndimage
 from torch.utils.data import Dataset
 
 from dataflux_pytorch import dataflux_mapstyle_dataset
+from demo.shared.image_segmentation import RandBalancedCrop
 
 
 class DatafluxPytTrain(Dataset):
@@ -66,17 +65,9 @@ class DatafluxPytTrain(Dataset):
         return len(self.images_dataset)
 
     def __getitem__(self, idx):
-        image = np.load(
-            io.BytesIO(
-                self.images_dataset[idx],
-            ),
-        )
+        image = np.load(io.BytesIO(self.images_dataset[idx], ), )
 
-        label = np.load(
-            io.BytesIO(
-                self.labels_dataset[idx],
-            ),
-        )
+        label = np.load(io.BytesIO(self.labels_dataset[idx], ), )
 
         data = {"image": image, "label": label}
         data = self.rand_crop(data)
@@ -96,73 +87,3 @@ class DatafluxPytTrain(Dataset):
             data = self.train_transforms(data)
             res.append((data["image"], data["label"]))
         return res
-
-
-class RandBalancedCrop:
-
-    def __init__(self, patch_size, oversampling):
-        self.patch_size = patch_size
-        self.oversampling = oversampling
-
-    def __call__(self, data):
-        image, label = data["image"], data["label"]
-        if random.random() < self.oversampling:
-            image, label, cords = self.rand_foreg_cropd(image, label)
-        else:
-            image, label, cords = self._rand_crop(image, label)
-        data.update({"image": image, "label": label})
-        return data
-
-    @staticmethod
-    def randrange(max_range):
-        return 0 if max_range == 0 else random.randrange(max_range)
-
-    def get_cords(self, cord, idx):
-        return cord[idx], cord[idx] + self.patch_size[idx]
-
-    def _rand_crop(self, image, label):
-        ranges = [s - p for s, p in zip(image.shape[1:], self.patch_size)]
-        cord = [self.randrange(x) for x in ranges]
-        low_x, high_x = self.get_cords(cord, 0)
-        low_y, high_y = self.get_cords(cord, 1)
-        low_z, high_z = self.get_cords(cord, 2)
-        image = image[:, low_x:high_x, low_y:high_y, low_z:high_z]
-        label = label[:, low_x:high_x, low_y:high_y, low_z:high_z]
-        return image, label, [low_x, high_x, low_y, high_y, low_z, high_z]
-
-    def rand_foreg_cropd(self, image, label):
-
-        def adjust(foreg_slice, patch_size, label, idx):
-            diff = patch_size[idx - 1] - (foreg_slice[idx].stop -
-                                          foreg_slice[idx].start)
-            sign = -1 if diff < 0 else 1
-            diff = abs(diff)
-            ladj = self.randrange(diff)
-            hadj = diff - ladj
-            low = max(0, foreg_slice[idx].start - sign * ladj)
-            high = min(label.shape[idx], foreg_slice[idx].stop + sign * hadj)
-            diff = patch_size[idx - 1] - (high - low)
-            if diff > 0 and low == 0:
-                high += diff
-            elif diff > 0:
-                low -= diff
-            return low, high
-
-        cl = np.random.choice(np.unique(label[label > 0]))
-        foreg_slices = scipy.ndimage.find_objects(
-            scipy.ndimage.measurements.label(label == cl)[0])
-        foreg_slices = [x for x in foreg_slices if x is not None]
-        slice_volumes = [
-            np.prod([s.stop - s.start for s in sl]) for sl in foreg_slices
-        ]
-        slice_idx = np.argsort(slice_volumes)[-2:]
-        foreg_slices = [foreg_slices[i] for i in slice_idx]
-        if not foreg_slices:
-            return self._rand_crop(image, label)
-        foreg_slice = foreg_slices[random.randrange(len(foreg_slices))]
-        low_x, high_x = adjust(foreg_slice, self.patch_size, label, 1)
-        low_y, high_y = adjust(foreg_slice, self.patch_size, label, 2)
-        low_z, high_z = adjust(foreg_slice, self.patch_size, label, 3)
-        image = image[:, low_x:high_x, low_y:high_y, low_z:high_z]
-        label = label[:, low_x:high_x, low_y:high_y, low_z:high_z]
-        return image, label, [low_x, high_x, low_y, high_y, low_z, high_z]
